# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Financial Control Plane API
  version: 1.0.0-alpha
  description: |
    Versioned REST API (PART 108). Commands are POST with an `Idempotency-Key` header (PART 36);
    queries are GET with cursor pagination. Errors are `application/problem+json` with a stable
    machine-readable `code` (PART 37). Every money field is a decimal string in USD (`"1234.56"`) or an
    exact base-unit integer string for asset quantities; no JSON numbers carry financial values.
    Timestamps are RFC 3339 UTC. Realtime updates are delivered over SSE and are never authoritative
    (PART 109): clients refetch canonical REST state after a stream gap.
servers:
  - url: /v1
...
security:
  - sessionCookie: []
paths:
...
  /auth/callback:
    get:
      tags: [auth]
      security: []
      summary: OIDC callback; establishes a server-side session cookie
      parameters:
        - { in: query, name: code, required: true, schema: { type: string } }
        - { in: query, name: state, required: true, schema: { type: string } }
      responses:
        "302": { description: Redirect to the application }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-cp_session
...
  responses:
    Problem:
      description: Structured error
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code:
          type: string
          description: Stable machine-readable code (INSUFFICIENT_BUYING_POWER, ACCOUNT_FROZEN, ASSET_RESTRICTED, VENUE_UNAVAILABLE, QUOTE_EXPIRED, RISK_MAX_POSITION, RISK_DAILY_LOSS, ELIGIBILITY_JURISDICTION, CAPABILITY_NOT_APPROVED, SUBMISSION_STATE_UNKNOWN, RECONCILIATION_REQUIRED, PROVIDER_UNAVAILABLE, STALE_MARKET_DATA, INVALID_IDEMPOTENCY_REUSE, KILL_SWITCH_ACTIVE, NO_VALID_PLAN, …)
        fields: { type: object, additionalProperties: true }
        request_id: { type: string }
...